+++ karma.conf.cjs
@@ -30,12 +30,6 @@ module.exports = function(config) {
       //mode: 'production',
       mode: 'development',
       devtool: 'inline-source-map',
-      node: {
-        Buffer: false,
-        process: false,
-        crypto: false,
-        setImmediate: false
-      }
     },
 
     // test results reporter to use