+++ tests/10-test.spec.js
@@ -3,140 +3,543 @@
  */
 'use strict';
 
-const chai = require('chai');
-chai.should();
+import chai from 'chai';
+import {Item, decodeDict} from 'structured-field-values';
+import * as httpSignatureHeader from '../lib/index.js';
+import {
+  signatureInputs,
+  signatures,
+  requests
+} from './mockData.js';
 
+const {HttpSignatureError} = httpSignatureHeader;
+chai.should();
 const {expect} = chai;
-const httpSignatureHeader = require('../lib');
-const HttpSignatureError = require('../lib/HttpSignatureError');
 
 describe('http-signature', () => {
+  describe('createSignatureHeader', () => {
+    it('should create a dictionary with one signature', () => {
+      const sigs = {sig1: new Uint8Array([1, 2, 3])};
+      const dict = httpSignatureHeader.createSignatureHeader(sigs);
+      expect(dict).to.be.a('string');
+      dict.should.be.equal('sig1=:AQID:');
+    });
+    it('should create a dictionary with multiple signatures', () => {
+      const sigs = {
+        sig1: new Uint8Array([1, 2, 3]),
+        sig2: new Uint8Array([4, 5, 6])
+      };
+      const dict = httpSignatureHeader.createSignatureHeader(sigs);
+      expect(dict).to.be.a('string');
+      dict.should.be.equal('sig1=:AQID:, sig2=:BAUG:');
+    });
+  });
+  describe('createSignatureInputHeader', () => {
+    const shouldBeAnFSDict = (actualDict, expectedDict) => {
+      expect(
+        actualDict, 'Expected dict to be a string').to.be.a('string');
+      actualDict.should.equal(expectedDict);
+      // should be able to parse it
+      const o = decodeDict(actualDict);
+      expect(o, 'expected decoded dict to be an object').to.be.an('object');
+    };
+    it('encodes a single signature input with no covered content', () => {
+      const dict = httpSignatureHeader.createSignatureInputHeader({
+        signatures: new Map([['sig1', {coveredContent: []}]]),
+        params: {alg: 'foo'}
+      });
+      shouldBeAnFSDict(dict, signatureInputs.empty);
+    });
+    it('encodes multiple signature inputs with no covered content', () => {
+      const dict = httpSignatureHeader.createSignatureInputHeader({
+        signatures: new Map([
+          ['sig1', {coveredContent: []}],
+          ['sig2', {coveredContent: []}]
+        ]),
+        params: {alg: 'foo'}
+      });
+      shouldBeAnFSDict(dict, signatureInputs.multipleEmpty);
+    });
+    it('encodes a single signature input with covered content' +
+      ' as strings', () => {
+      const dict = httpSignatureHeader.createSignatureInputHeader({
+        signatures: new Map([['sig1', {coveredContent: ['one']}]]),
+        params: {alg: 'foo'}
+      });
+      shouldBeAnFSDict(dict, signatureInputs.one);
+    });
+    it('encodes multiple signature inputs with covered content' +
+      ' as strings', () => {
+      const dict = httpSignatureHeader.createSignatureInputHeader({
+        signatures: new Map([
+          ['sig1', {coveredContent: ['one']}],
+          ['sig2', {coveredContent: ['two']}]
+        ]),
+        params: {alg: 'foo'}
+      });
+      shouldBeAnFSDict(dict, signatureInputs.oneMultiple);
+    });
+    it('encodes a single signature input with covered content' +
+      ' as Items with no params', () => {
+      const dict = httpSignatureHeader.createSignatureInputHeader({
+        signatures: new Map([['sig1', {coveredContent: [{value: 'one'}]}]]),
+        params: {alg: 'foo'}
+      });
+      shouldBeAnFSDict(dict, signatureInputs.one);
+    });
+    it('encodes multiple signature inputs with covered content' +
+      ' as Items with no params', () => {
+      const dict = httpSignatureHeader.createSignatureInputHeader({
+        signatures: new Map([
+          ['sig1', {coveredContent: [{value: 'one'}]}],
+          ['sig2', {coveredContent: [{value: 'two'}]}]
+        ]),
+        params: {alg: 'foo'}
+      });
+      shouldBeAnFSDict(dict, signatureInputs.oneMultiple);
+    });
+    it('encodes a single signature input with covered content' +
+      ' as Items with prefix params', () => {
+      const dict = httpSignatureHeader.createSignatureInputHeader({
+        signatures: new Map([
+          ['sig1', {coveredContent: [{value: 'one', params: {prefix: 3}}]}]
+        ]),
+        params: {alg: 'foo'}
+      });
+      shouldBeAnFSDict(dict, signatureInputs.onePrefix);
+    });
+    it('encodes multiple signature inputs with covered content' +
+      ' as Items with prefix params', () => {
+      const dict = httpSignatureHeader.createSignatureInputHeader({
+        signatures: new Map([
+          ['sig1', {coveredContent: [{value: 'one', params: {prefix: 1}}]}],
+          ['sig2', {coveredContent: [{value: 'two', params: {prefix: 1}}]}]
+        ]),
+        params: {alg: 'foo'}
+      });
+      shouldBeAnFSDict(dict, signatureInputs.multiplePrefix);
+    });
+    it('encodes a single signature input with covered content' +
+      ' as Items with key params', () => {
+      const dict = httpSignatureHeader.createSignatureInputHeader({
+        signatures: new Map([
+          ['sig1', {coveredContent: [{value: 'one', params: {key: 'one'}}]}]
+        ]),
+        params: {alg: 'foo'}
+      });
+      shouldBeAnFSDict(dict, signatureInputs.oneKey);
+    });
+    it('encodes multiple signature inputs with covered content' +
+      ' as Items with key params', () => {
+      const dict = httpSignatureHeader.createSignatureInputHeader({
+        signatures: new Map([
+          ['sig1', {coveredContent: [{value: 'one', params: {key: 'one'}}]}],
+          ['sig2', {coveredContent: [{value: 'two', params: {key: 'two'}}]}]
+        ]),
+        params: {alg: 'foo'}
+      });
+      shouldBeAnFSDict(dict, signatureInputs.multipleKey);
+    });
 
-  describe('createSignatureString API', () => {
+  });
+  describe('createSignatureInputString', () => {
     it('uses `date` header if specified', () => {
       const date = new Date().toUTCString();
-      const requestOptions = {
+      const httpMessage = {
         headers: {date},
         method: 'GET',
         url: 'https://example.com',
       };
-      const stringToSign = httpSignatureHeader.createSignatureString(
-        {includeHeaders: ['date'], requestOptions});
-      stringToSign.should.equal(`date: ${date}`);
+      const {sig1} = decodeDict(signatureInputs.date);
+      const stringToSign = httpSignatureHeader.createSignatureInputString({
+        signatureInput: sig1,
+        httpMessage
+      });
+      stringToSign.should.equal(`"date": ${date}\n` +
+        '"@signature-params": ("date");keyid="foo";alg="bar"');
     });
-    it('properly encodes `(request-target)` with root path', () => {
+    it('properly encodes `@request-target` with root path', () => {
       const date = new Date().toUTCString();
-      const requestOptions = {
+      const httpMessage = {
         headers: {date},
         method: 'GET',
         url: 'https://example.com',
       };
-      const stringToSign = httpSignatureHeader.createSignatureString(
-        {includeHeaders: ['date', '(request-target)'], requestOptions});
-      stringToSign.should.equal(`date: ${date}\n(request-target): get /`);
+      const {sig1} = decodeDict(signatureInputs.requestTarget);
+      const stringToSign = httpSignatureHeader.createSignatureInputString({
+        signatureInput: sig1,
+        httpMessage
+      });
+      stringToSign.should.equal(`"date": ${date}\n"@request-target": get /` +
+        '\n"@signature-params": ("date" "@request-target");keyid="foo"' +
+        ';alg="bar"');
     });
-    it('properly encodes `(request-target)` with a path', () => {
+    it('properly encodes `@request-target` with a path', () => {
       const date = new Date().toUTCString();
-      const requestOptions = {
+      const httpMessage = {
         headers: {date},
         method: 'GET',
         url: 'https://example.com/1/2/3',
       };
-      const stringToSign = httpSignatureHeader.createSignatureString(
-        {includeHeaders: ['date', '(request-target)'], requestOptions});
-      stringToSign.should.equal(`date: ${date}\n(request-target): get /1/2/3`);
+      const {sig1} = decodeDict(signatureInputs.requestTarget);
+      const stringToSign = httpSignatureHeader.createSignatureInputString({
+        signatureInput: sig1,
+        httpMessage
+      });
+      stringToSign.should.equal(
+        `"date": ${date}\n"@request-target": get /1/2/3` +
+        '\n"@signature-params": ("date" "@request-target");keyid="foo"' +
+        ';alg="bar"'
+      );
     });
-    it('properly encodes `(request-target)` with a relative path', () => {
+    it('properly encodes `@request-target` with a path using ' +
+      'coveredContent & params', () => {
       const date = new Date().toUTCString();
-      const requestOptions = {
+      const httpMessage = {
+        headers: {date},
+        method: 'GET',
+        url: 'https://example.com/1/2/3',
+      };
+      const stringToSign = httpSignatureHeader.createSignatureInputString({
+        coveredContent: ['date', '@request-target'],
+        params: {keyid: 'foo', alg: 'bar'},
+        httpMessage
+      });
+      stringToSign.should.equal(
+        `"date": ${date}\n"@request-target": get /1/2/3` +
+        '\n"@signature-params": ("date" "@request-target");keyid="foo"' +
+        ';alg="bar"'
+      );
+    });
+
+    it('properly encodes `@request-target` with a relative path', () => {
+      const date = new Date().toUTCString();
+      const httpMessage = {
         headers: {date},
         method: 'GET',
         url: '/relative/path',
       };
-      const stringToSign = httpSignatureHeader.createSignatureString(
-        {includeHeaders: ['date', '(request-target)'], requestOptions});
+      const {sig1} = decodeDict(signatureInputs.requestTarget);
+      const stringToSign = httpSignatureHeader.createSignatureInputString({
+        signatureInput: sig1,
+        httpMessage
+      });
       stringToSign.should.equal(
-        `date: ${date}\n(request-target): get /relative/path`);
+        `"date": ${date}\n"@request-target": get /relative/path` +
+        '\n"@signature-params": ("date" "@request-target");keyid="foo"' +
+        ';alg="bar"'
+      );
     });
-    it('properly encodes `(request-target)` with post method', () => {
+    it('properly encodes `@request-target` with post method', () => {
       const date = new Date().toUTCString();
-      const requestOptions = {
+      const httpMessage = {
         headers: {date},
         method: 'POST',
         url: 'https://example.com',
       };
-      const stringToSign = httpSignatureHeader.createSignatureString(
-        {includeHeaders: ['date', '(request-target)'], requestOptions});
-      stringToSign.should.equal(`date: ${date}\n(request-target): post /`);
+      const {sig1} = decodeDict(signatureInputs.requestTarget);
+      const stringToSign = httpSignatureHeader.createSignatureInputString({
+        signatureInput: sig1,
+        httpMessage
+      });
+      stringToSign.should.equal(
+        `"date": ${date}\n"@request-target": post /` +
+        '\n"@signature-params": ("date" "@request-target");keyid="foo"' +
+        ';alg="bar"'
+      );
     });
 
     it('properly encodes `host`', () => {
       const date = new Date().toUTCString();
-      const requestOptions = {
-        headers: {date},
+      const httpMessage = {
+        headers: {date, host: 'example.com'},
         method: 'GET',
         url: 'https://example.com/1/2/3',
       };
-      const stringToSign = httpSignatureHeader.createSignatureString(
-        {includeHeaders: ['host', 'date', '(request-target)'], requestOptions});
+      const {sig1} = decodeDict(signatureInputs.host);
+      const stringToSign = httpSignatureHeader.createSignatureInputString({
+        signatureInput: sig1,
+        httpMessage
+      });
       stringToSign.should.equal(
-        `host: example.com\ndate: ${date}\n(request-target): get /1/2/3`);
+        `"host": example.com\n"date": ${date}\n"@request-target": get /1/2/3` +
+        '\n"@signature-params": ("host" "date" "@request-target")' +
+        ';keyid="foo";alg="bar"'
+      );
     });
-
     it('properly encodes `host` with a port', () => {
       const date = new Date().toUTCString();
-      const requestOptions = {
-        headers: {date},
+      const httpMessage = {
+        headers: {date, host: 'example.com:18443'},
         method: 'GET',
         url: 'https://example.com:18443/1/2/3',
       };
-      const stringToSign = httpSignatureHeader.createSignatureString(
-        {includeHeaders: ['host', 'date', '(request-target)'], requestOptions});
+      const {sig1} = decodeDict(signatureInputs.host);
+      const stringToSign = httpSignatureHeader.createSignatureInputString({
+        signatureInput: sig1,
+        httpMessage
+      });
       stringToSign.should.equal(
-        `host: example.com:18443\ndate: ${date}\n(request-target): get /1/2/3`);
+        `"host": example.com:18443\n"date": ${date}\n"@request-target": get ` +
+        '/1/2/3\n"@signature-params": ("host" "date" "@request-target")' +
+        ';keyid="foo";alg="bar"'
+      );
     });
-
     it('properly encodes a header with a zero-length value', () => {
       const date = new Date().toUTCString();
       const zero = '';
-      const requestOptions = {
+      const httpMessage = {
         headers: {date, zero},
         method: 'GET',
         url: 'https://example.com:18443/1/2/3',
       };
-      const stringToSign = httpSignatureHeader.createSignatureString(
-        {includeHeaders: ['host', 'date', 'zero'], requestOptions});
+      const {sig1} = decodeDict(
+        'sig1=("date" "zero");keyid="foo";alg="bar"');
+      const stringToSign = httpSignatureHeader.createSignatureInputString({
+        signatureInput: sig1,
+        httpMessage
+      });
       stringToSign.should.equal(
-        `host: example.com:18443\ndate: ${date}\nzero: `);
+        `"date": ${date}\n"zero": ` +
+        '\n"@signature-params": ("date" "zero")' +
+        ';keyid="foo";alg="bar"'
+      );
     });
 
     it('properly encodes a header that\'s value is all white spaces', () => {
       const date = new Date().toUTCString();
       const zero = '  ';
-      const requestOptions = {
+      const httpMessage = {
         headers: {date, zero},
         method: 'GET',
         url: 'https://example.com:18443/1/2/3',
       };
-      const stringToSign = httpSignatureHeader.createSignatureString(
-        {includeHeaders: ['host', 'date', 'zero'], requestOptions});
+      const {sig1} = decodeDict(signatureInputs.zero);
+      const stringToSign = httpSignatureHeader.createSignatureInputString({
+        signatureInput: sig1,
+        httpMessage
+      });
       stringToSign.should.equal(
-        `host: example.com:18443\ndate: ${date}\nzero: `);
+        `"date": ${date}\n"zero": ` +
+        '\n"@signature-params": ("date" "zero")' +
+        ';keyid="foo";alg="bar"'
+      );
     });
-
     it('properly encodes a header with multiple values', () => {
       const date = new Date().toUTCString();
       const multiple = 'true, false';
-      const requestOptions = {
+      const httpMessage = {
         headers: {date, multiple},
         method: 'GET',
         url: 'https://example.com:18443/1/2/3',
       };
-      const stringToSign = httpSignatureHeader.createSignatureString(
-        {includeHeaders: ['host', 'date', 'multiple'], requestOptions});
+      const {sig1} = decodeDict(signatureInputs.multipleValues);
+      const stringToSign = httpSignatureHeader.createSignatureInputString({
+        signatureInput: sig1,
+        httpMessage
+      });
       stringToSign.should.equal(
-        `host: example.com:18443\ndate: ${date}\nmultiple: true, false`);
+        `"date": ${date}\n"multiple": true, false` +
+        '\n"@signature-params": ("date" "multiple")' +
+        ';keyid="foo";alg="bar"'
+      );
     });
+    it('properly encodes a header with a dictionary value', () => {
+      const date = new Date().toUTCString();
+      const dictionary = 'test="foo"';
+      const httpMessage = {
+        headers: {date, dictionary},
+        method: 'GET',
+        url: 'https://example.com:18443/1/2/3',
+      };
+      const {sig1} = decodeDict(signatureInputs.dict);
+      const stringToSign = httpSignatureHeader.createSignatureInputString({
+        signatureInput: sig1,
+        httpMessage
+      });
+      stringToSign.should.equal(
+        `"date": ${date}\n"dictionary": foo` +
+        '\n"@signature-params": ("date" "dictionary";key="test")' +
+        ';keyid="foo";alg="bar"'
+      );
+    });
+    it('throws when a content identifier has a key param but header ' +
+      'is not a dictionary', () => {
+      const date = new Date().toUTCString();
+      const dictionary = '("foo")';
+      const httpMessage = {
+        headers: {date, dictionary},
+        method: 'GET',
+        url: 'https://example.com:18443/1/2/3',
+      };
+      const {sig1} = decodeDict(signatureInputs.dict);
+      expect(() => httpSignatureHeader.createSignatureInputString(
+        {signatureInput: sig1, httpMessage}))
+        .to.throw(Error, /failed to parse/i);
+    });
+    it('throws when a content identifier has a key param but the key ' +
+      'is not found in the dictionary', () => {
+      const date = new Date().toUTCString();
+      const dictionary = 'test="foo"';
+      const httpMessage = {
+        headers: {date, dictionary},
+        method: 'GET',
+        url: 'https://example.com:18443/1/2/3',
+      };
+      const {sig1} = decodeDict(
+        'sig1=("date" "dictionary";key="bar");keyid="foo";alg="bar"');
+      expect(() => httpSignatureHeader.createSignatureInputString(
+        {signatureInput: sig1, httpMessage}))
+        .to.throw(HttpSignatureError, /Failed to find key/i);
+    });
+    it('properly encodes a header with a list value', () => {
+      const date = new Date().toUTCString();
+      const list = '("foo" "bar")';
+      const httpMessage = {
+        headers: {date, list},
+        method: 'GET',
+        url: 'https://example.com:18443/1/2/3',
+      };
+      const {sig1} = decodeDict(
+        'sig1=("date" "list";prefix=1);keyid="foo";alg="bar"');
+      const stringToSign = httpSignatureHeader.createSignatureInputString({
+        signatureInput: sig1,
+        httpMessage
+      });
+      stringToSign.should.equal(
+        `"date": ${date}\n"list": (foo)` +
+        '\n"@signature-params": ("date" "list";prefix=1)' +
+        ';keyid="foo";alg="bar"'
+      );
+    });
+    it('properly encodes when a prefix parameter should return' +
+      ' multiple values', () => {
+      const date = new Date().toUTCString();
+      const list = '(1 2 3 4 5 6)';
+      const httpMessage = {
+        headers: {date, list},
+        method: 'GET',
+        url: 'https://example.com:18443/1/2/3',
+      };
+      const {sig1} = decodeDict(
+        'sig1=("date" "list";prefix=3);keyid="foo";alg="bar"');
+      const stringToSign = httpSignatureHeader.createSignatureInputString({
+        signatureInput: sig1,
+        httpMessage
+      });
+      stringToSign.should.equal(
+        `"date": ${date}\n"list": (1, 2, 3)` +
+        '\n"@signature-params": ("date" "list";prefix=3)' +
+        ';keyid="foo";alg="bar"'
+      );
+    });
+
+    it('properly encodes when a prefix parameter should return' +
+      ' no values', () => {
+      const date = new Date().toUTCString();
+      const list = '(1 2 3 4 5 6)';
+      const httpMessage = {
+        headers: {date, list},
+        method: 'GET',
+        url: 'https://example.com:18443/1/2/3',
+      };
+      const {sig1} = decodeDict(
+        'sig1=("date" "list";prefix=0);keyid="foo";alg="bar"');
+      const stringToSign = httpSignatureHeader.createSignatureInputString({
+        signatureInput: sig1,
+        httpMessage
+      });
+      stringToSign.should.equal(
+        `"date": ${date}\n"list": ()` +
+        '\n"@signature-params": ("date" "list";prefix=0)' +
+        ';keyid="foo";alg="bar"'
+      );
+    });
+
+    it('throws when a content identifier has a prefix param but ' +
+      'the header is not a list', () => {
+      const date = new Date().toUTCString();
+      const list = 'test="foo"';
+      const httpMessage = {
+        headers: {date, list},
+        method: 'GET',
+        url: 'https://example.com:18443/1/2/3',
+      };
+      const {sig1} = decodeDict(
+        'sig1=("date" "list";prefix=0);keyid="foo";alg="bar"');
+      expect(() => httpSignatureHeader.createSignatureInputString(
+        {signatureInput: sig1, httpMessage}))
+        .to.throw(Error, /Failed to parse/i);
+    });
+    it('throws when a content identifier has a prefix param but ' +
+      'the prefix is out of bounds', () => {
+      const date = new Date().toUTCString();
+      const list = '("foo" "bar")';
+      const httpMessage = {
+        headers: {date, list},
+        method: 'GET',
+        url: 'https://example.com:18443/1/2/3',
+      };
+      const {sig1} = decodeDict(
+        'sig1=("date" "list";prefix=3);keyid="foo";alg="bar"');
+      expect(() => httpSignatureHeader.createSignatureInputString(
+        {signatureInput: sig1, httpMessage}))
+        .to.throw(Error, /out of bounds/i);
+    });
+
+    it('properly encodes using header parameter order', () => {
+      const date = new Date().toUTCString();
+      const httpMessage = {
+        headers: {date, host: 'example.com'},
+        method: 'GET',
+        url: 'https://example.com/1/2/3',
+      };
+      // 2 different signatures with the same parameters
+      // in a different order
+      const {sig1} = decodeDict(
+        'sig1=("host" "date" "@request-target");keyid="foo";alg="bar"');
+      const stringToSign1 = httpSignatureHeader.createSignatureInputString({
+        signatureInput: sig1,
+        httpMessage
+      });
+      // the header parameters have the same value, but are in a different order
+      const {sig2} = decodeDict(
+        'sig2=("date" "host" "@request-target");keyid="foo";alg="bar"');
+      const stringToSign2 = httpSignatureHeader.createSignatureInputString({
+        signatureInput: sig2,
+        httpMessage
+      });
+      stringToSign2.should.not.equal(stringToSign1);
+    });
+
+    it('throws when an unknown header is specified', () => {
+      const date = new Date().toUTCString();
+      const httpMessage = {
+        headers: {date},
+        method: 'GET',
+        url: 'https://example.com:18443/1/2/3',
+      };
+      const {sig1} = decodeDict(
+        'sig1=("foo" "date");keyid="foo";alg="bar"');
+      expect(() => httpSignatureHeader.createSignatureInputString(
+        {signatureInput: sig1, httpMessage}))
+        .to.throw(HttpSignatureError, `Failed to find header field name "foo"`);
+    });
+    it('throws when an invalid header is specified', () => {
+      const httpMessage = {
+        headers: {'(bad)': true},
+        method: 'GET',
+        url: 'https://example.com:18443/1/2/3',
+      };
+      const {sig1} = decodeDict(
+        'sig1=("(bad)");keyid="foo";alg="bar"');
+      expect(() => httpSignatureHeader.createSignatureInputString(
+        {signatureInput: sig1, httpMessage}))
+        .to.throw(HttpSignatureError, /Illegal header "[A-z\(\)]+"/i);
+    });
+  });
+  /*
+  describe.skip('createSignatureString API', () => {
 
     it('properly encodes `(key-id)` with an iri', () => {
       const iri = 'https://example.com/key.pub';
@@ -232,228 +635,57 @@ describe('http-signature', () => {
       });
 
     }
-
-    it('properly encodes a header with multiple values', () => {
-      const date = new Date().toUTCString();
-      const requestOptions = {
-        headers: {date, 'x-custom': ['val1', 'val2']},
-        method: 'GET',
-        url: 'https://example.com/1/2/3',
-      };
-      const stringToSign = httpSignatureHeader.createSignatureString(
-        {includeHeaders: ['x-custom', 'host', 'date', '(request-target)'],
-          requestOptions});
-      stringToSign.should.equal(
-        `x-custom: val1, val2\nhost: example.com\ndate: ` +
-        `${date}\n(request-target): get /1/2/3`);
+  });
+*/
+  describe('parseRequest API', function() {
+    const shouldBeParsedSignature = actualSig => {
+      expect(actualSig).to.be.an('object');
+      expect(actualSig.errors).to.be.an('Array');
+      expect(
+        actualSig.signatureInput,
+        'expected "signatureInput" to be a structured field Item'
+      ).to.be.an.instanceOf(Item);
+      expect(
+        actualSig.signature,
+        'expected "signature" to be a structured field Item'
+      ).to.be.an.instanceOf(Item);
+      expect(actualSig.signingString).to.be.a('string');
+    };
+    it('should parse a Signature with no covered content', () => {
+      const request = {...requests.exampleOne};
+      request.headers['Signature-Input'] = signatureInputs.exampleOne;
+      request.headers.Signature = signatures.exampleOne;
+      const result = httpSignatureHeader.parseRequest(request);
+      expect(result).to.be.a('Map');
+      const sig1 = result.get('sig1');
+      shouldBeParsedSignature(sig1);
+      sig1.errors.length.should.equal(0);
+      sig1.signingString.should.equal('"@signature-params": ()' +
+      ';created=1618884475;keyid="test-key-rsa-pss";alg="rsa-pss-sha512"');
     });
-
-    it('properly encodes using header parameter order', () => {
-      const date = new Date().toUTCString();
-      const requestOptions = {
-        headers: {date},
-        method: 'GET',
-        url: 'https://example.com/1/2/3',
-      };
-      // 2 different signatures with the same parameters
-      // in a different order
-      const firstString = httpSignatureHeader.createSignatureString(
-        {includeHeaders: ['host', 'date', '(request-target)'],
-          requestOptions});
-      firstString.should.equal(
-        `host: example.com\ndate: ${date}\n` +
-        `(request-target): get /1/2/3`);
-      // the header parameters have the same value, but are in a different order
-      const secondString = httpSignatureHeader.createSignatureString(
-        {includeHeaders: ['(request-target)', 'host', 'date'],
-          requestOptions});
-      secondString.should.not.equal(firstString);
-      secondString.should.equal(
-        `(request-target): get /1/2/3\nhost: example.com\ndate: ` +
-        `${date}`);
-    });
-
-    it('throws when an unknown header is specified', () => {
-      const date = new Date().toUTCString();
-      const requestOptions = {
-        headers: {date},
-        method: 'GET',
-        url: 'https://example.com:18443/1/2/3',
-      };
-      expect(() => httpSignatureHeader.createSignatureString(
-        {includeHeaders: ['foo', 'date'], requestOptions}))
-        .to.throw(HttpSignatureError, /foo was not found/);
-    });
-    it('throws when an invalid header is specified', () => {
-      const requestOptions = {
-        headers: {'(bad)': true},
-        method: 'GET',
-        url: 'https://example.com:18443/1/2/3',
-      };
-      expect(() => httpSignatureHeader.createSignatureString(
-        {includeHeaders: ['(bad)'], requestOptions}))
-        .to.throw(HttpSignatureError, /Illegal header "[A-z\(\)]+"/i);
+    it('should parse a Signature with covered content', () => {
+      const request = {...requests.exampleOne};
+      request.headers['Signature-Input'] = signatureInputs.exampleTwo;
+      request.headers.Signature = signatures.exampleTwo;
+      const result = httpSignatureHeader.parseRequest(request);
+      expect(result).to.be.a('Map');
+      const sig1 = result.get('sig1');
+      shouldBeParsedSignature(sig1);
+      sig1.errors.length.should.equal(0);
+      sig1.signingString.should.equal(
+        '"@request-target": post /foo?param=value&pet=dog\n' +
+        '"host": example.com\n' +
+        '"date": Tue, 20 Apr 2021 02:07:55 GMT\n' +
+        '"content-type": application/json\n' +
+        '"digest": SHA-256=X48E9qOokqqrvdts8nOJRJN3OWDUoyWxBf7kbu9DBPE=\n' +
+        '"content-length": 18\n' +
+        '"@signature-params": ("@request-target"' +
+        ' "host" "date" "content-type" "digest" "content-length")' +
+        ';created=1618884475;keyid="test-key-rsa-pss"'
+      );
     });
   });
-
-  describe('createAuthzHeader API', () => {
-    it('header with an algorithm and one specified header', () => {
-      const authz = httpSignatureHeader.createAuthzHeader({
-        algorithm: 'rsa-sha256',
-        includeHeaders: ['date'],
-        keyId: 'https://example.com/key/1',
-        signature: 'mockSignature'
-      });
-      authz.should.be.a('string');
-      authz.should.equal('Signature keyId="https://example.com/key/1",' +
-        'algorithm="rsa-sha256",headers="date",signature="mockSignature"');
-    });
-    it('header with an algorithm and two specified headers', () => {
-      const authz = httpSignatureHeader.createAuthzHeader({
-        algorithm: 'rsa-sha256',
-        includeHeaders: ['date', 'host'],
-        keyId: 'https://example.com/key/1',
-        signature: 'mockSignature'
-      });
-      authz.should.be.a('string');
-      authz.should.equal('Signature keyId="https://example.com/key/1",' +
-        'algorithm="rsa-sha256",headers="date host",signature="mockSignature"');
-    });
-    it('header with an algorithm and three specified headers', () => {
-      const authz = httpSignatureHeader.createAuthzHeader({
-        algorithm: 'rsa-sha256',
-        includeHeaders: ['date', 'host', '(request-target)'],
-        keyId: 'https://example.com/key/1',
-        signature: 'mockSignature'
-      });
-      authz.should.be.a('string');
-      authz.should.equal('Signature keyId="https://example.com/key/1",' +
-        'algorithm="rsa-sha256",headers="date host (request-target)",' +
-        'signature="mockSignature"');
-    });
-    it('header without an algorithm and one specified header', () => {
-      const authz = httpSignatureHeader.createAuthzHeader({
-        includeHeaders: ['date'],
-        keyId: 'https://example.com/key/1',
-        signature: 'mockSignature'
-      });
-      authz.should.be.a('string');
-      authz.should.equal('Signature keyId="https://example.com/key/1",' +
-        'headers="date",signature="mockSignature"');
-    });
-    it('header without an algorithm and two specified headers', () => {
-      const authz = httpSignatureHeader.createAuthzHeader({
-        includeHeaders: ['date', 'host'],
-        keyId: 'https://example.com/key/1',
-        signature: 'mockSignature'
-      });
-      authz.should.be.a('string');
-      authz.should.equal('Signature keyId="https://example.com/key/1",' +
-        'headers="date host",signature="mockSignature"');
-    });
-    it('header without an algorithm and three specified headers', () => {
-      const authz = httpSignatureHeader.createAuthzHeader({
-        includeHeaders: ['date', 'host', '(request-target)'],
-        keyId: 'https://example.com/key/1',
-        signature: 'mockSignature'
-      });
-      authz.should.be.a('string');
-      authz.should.equal('Signature keyId="https://example.com/key/1",' +
-        'headers="date host (request-target)",' +
-        'signature="mockSignature"');
-    });
-    it('header with integer (created)', () => {
-      const created = Math.floor(Date.now() / 1000);
-      const authz = httpSignatureHeader.createAuthzHeader({
-        includeHeaders: ['date', 'host', '(request-target)', '(created)'],
-        keyId: 'https://example.com/key/1',
-        signature: 'mockSignature',
-        created
-      });
-      authz.should.be.a('string');
-      authz.should.equal('Signature keyId="https://example.com/key/1",' +
-        'headers="date host (request-target) (created)",' +
-        `signature="mockSignature",created="${created}"`);
-    });
-    it('header with string (created)', () => {
-      const created = String(Math.floor(Date.now() / 1000));
-      const authz = httpSignatureHeader.createAuthzHeader({
-        includeHeaders: ['date', 'host', '(request-target)', '(created)'],
-        keyId: 'https://example.com/key/1',
-        signature: 'mockSignature',
-        created
-      });
-      authz.should.be.a('string');
-      authz.should.equal('Signature keyId="https://example.com/key/1",' +
-        'headers="date host (request-target) (created)",' +
-        `signature="mockSignature",created="${created}"`);
-    });
-    it('should throw if (created) is not a unix timestamp', () => {
-      const created = 'invalid-date-time';
-      let error = null;
-      let result = null;
-      try {
-        result = httpSignatureHeader.createAuthzHeader({
-          includeHeaders: ['date', 'host', '(request-target)', '(created)'],
-          keyId: 'https://example.com/key/1',
-          signature: 'mockSignature',
-          created
-        });
-      } catch(e) {
-        error = e;
-      }
-      expect(result).to.be.null;
-      expect(error).to.not.be.null;
-      error.message.should.equal(
-        '"created" must be a UNIX timestamp or JavaScript Date.');
-    });
-    it('header with integer (expires)', () => {
-      const expires = Math.floor(Date.now() / 1000);
-      const authz = httpSignatureHeader.createAuthzHeader({
-        includeHeaders: ['date', 'host', '(request-target)', '(expires)'],
-        keyId: 'https://example.com/key/1',
-        signature: 'mockSignature',
-        expires
-      });
-      authz.should.be.a('string');
-      authz.should.equal('Signature keyId="https://example.com/key/1",' +
-        'headers="date host (request-target) (expires)",' +
-        `signature="mockSignature",expires="${expires}"`);
-    });
-    it('header with string (expires)', () => {
-      const expires = String(Math.floor(Date.now() / 1000));
-      const authz = httpSignatureHeader.createAuthzHeader({
-        includeHeaders: ['date', 'host', '(request-target)', '(expires)'],
-        keyId: 'https://example.com/key/1',
-        signature: 'mockSignature',
-        expires
-      });
-      authz.should.be.a('string');
-      authz.should.equal('Signature keyId="https://example.com/key/1",' +
-        'headers="date host (request-target) (expires)",' +
-        `signature="mockSignature",expires="${expires}"`);
-    });
-    it('should throw if (expires) is not a unix timestamp', () => {
-      const expires = 'invalid-date-time';
-      let error = null;
-      let result = null;
-      try {
-        result = httpSignatureHeader.createAuthzHeader({
-          includeHeaders: ['date', 'host', '(request-target)', '(expires)'],
-          keyId: 'https://example.com/key/1',
-          signature: 'mockSignature',
-          expires
-        });
-      } catch(e) {
-        error = e;
-      }
-      expect(result).to.be.null;
-      expect(error).to.not.be.null;
-      error.message.should.equal(
-        '"expires" must be a UNIX timestamp or JavaScript Date.');
-    });
-
-  });
+/*
   describe('parseRequest API', function() {
     // takes the result of parseRequest and tests it
     const shouldBeParsed = parsed => {
@@ -1027,4 +1259,5 @@ describe('http-signature', () => {
       error.message.should.equal('(expires) was not a signed header');
     });
   });
+*/
 });